Few additions & facing enemy on hit
Player will now face the enemy that attacks it on hit. Also added a macro library and parent class called damageable objects (BP). If you add any blueprints that can be damaged in the game make sure to make them a child of this blueprint and use the template on hit in said blueprint