Anubis boss fight start
Anubis has been added to level Trial of combat. He can attack (without weapon so it doesn't really work). Fire a fireball (need to make it AOE). He cycles attack on random at the moment but eventually he will detect the distance from the player and decide to use an attack from that