3 Gate Mechanics
I have made a sliding gate, a pressure pad-based gate and a disappearing one.
These have been set up in a way that they can be modified whenever needed. In terms of polish (appearance, VFX and SFX etc).
The gates work in the way that I intended them to.

(The rest of it is mentioned by GitHub).