Some more Gate Mechanics
I have added extra folders within my Gate mechanics, for animations (using the Level Sequencers).
One of them is for a moving platform; I was asked to start that. But it follows a similar approach to some of my moving gates/ doors, so that was fine.
Unfortunately, the sequencers will only work when the gates have been placed in final locations, but once they have I can update those elements.
I have also added in a rotating door (gate), in case those are wanted or needed.